bug: fix a major memory leak
This one is on me - I'd missed that I added a Kleene plus after a
Kleene star, causing the regular expression to never progress on
some platforms (and therefore allocating basically all of the
available memory on the system, and crashing the process).

Simplifying this matcher rule gets rid of the excessively greedy
matcher and fixes the memory leak.